fastapi_backend/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn==0.30.6
numpy
pillow
pymupdf
rapidocr-onnxruntime
faster-whisper
